% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtEst.R
\name{beta2Est}
\alias{beta2Est}
\title{Calculate sdt-beta2 from hit and false alarm rates.}
\usage{
beta2Est(h = NULL, fa = NULL, dp = NULL, c = NULL)
}
\arguments{
\item{h}{The hit rate (as a proportion).}

\item{fa}{The false alarm rate (as a proportion).}

\item{dp}{d' from signal detection. Optional (see details).}

\item{c}{Criterion from signal detection. Optional (see details).}
}
\value{
The c parameter from SDT.
}
\description{
Calculate sdt-beta2 from hit and false alarm rates.
}
\details{
Either the hit and false alarm rates can be specified, or d' and 
criterion. If either d' or criterion is specified, then both must be
provided. If either hit rate or false alarm rate is specified, then both
must be provided. Note that if hit rate and false alarm rate
are provided, they are first used to calculate d' and c. Thus, it is
faster to provide d' and c if they are available.
}
\examples{
beta2Est(.71, .32)

}
