% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtEst.R
\name{betaEst}
\alias{betaEst}
\title{Calculate sdt-beta from hit and false alarm rates.}
\usage{
betaEst(h, fa)
}
\arguments{
\item{h}{The hit rate (as a proportion).}

\item{fa}{The false alarm rate (as a proportion).}
}
\value{
The beta parameter from SDT.
}
\description{
Calculate sdt-beta from hit and false alarm rates.
}
\examples{
betaEst(.71, .32)

}
