% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtEst.R
\name{cEst}
\alias{cEst}
\title{Calculate criterion from hit and false alarm rates.}
\usage{
cEst(h, fa)
}
\arguments{
\item{h}{The hit rate (as a proportion).}

\item{fa}{The false alarm rate (as a proportion).}
}
\value{
The c parameter from SDT.
}
\description{
Calculate criterion from hit and false alarm rates.
}
\examples{
cEst(.71, .32)

}
