% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtEst.R
\name{dpEst}
\alias{dpEst}
\title{Calculate d' from hit and false alarm rates.}
\usage{
dpEst(h, fa)
}
\arguments{
\item{h}{The hit rate (as a proportion).}

\item{fa}{The false alarm rate (as a proportion).}
}
\value{
The d' parameter from SDT.
}
\description{
Calculate d' from hit and false alarm rates.
}
\examples{
dpEst(.75, .31)

}
