% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSdt.R
\name{plotSDT}
\alias{plotSDT}
\title{Plot of an SDT analysis}
\usage{
plotSDT(dp = 2, crit = NULL, xlim = c(-4, 4), noiseLab = "",
  signalLab = "", xlab = "evidence", ylab = "", main = "", dpLab = NA,
  noiseCol = "darkblue", signalCol = "coral3", critCol = "black",
  critLty = 1, critLwd = 1, critLab = NA, noiseLty = 1, signalLty = 1,
  noiseLwd = 1, signalLwd = 1)
}
\arguments{
\item{dp}{The d-prime value to show.}

\item{crit}{The criterion value.}

\item{xlim}{The limits of the x-axis.}

\item{noiseLab}{The label to apply to the noise distribution, if any.}

\item{signalLab}{The label to apply to the signal distribution, if any.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{main}{The plot title.}

\item{dpLab}{The label to apply to the d-prime representation.}

\item{noiseCol}{The color of the noise distribution.}

\item{signalCol}{The color of the signal distribution.}

\item{critCol}{The color of the criterion line.}

\item{critLty}{The type of line to draw for the criterion.}

\item{critLwd}{Criterion line width.}

\item{critLab}{The criterion label to apply.}

\item{noiseLty}{The type of line to draw for the noise distribution.}

\item{signalLty}{The type of line to draw for the signal distribution.}

\item{noiseLwd}{The width of the noise distribution line.}

\item{signalLwd}{the width of the signal distribution line.}
}
\description{
Plot of an SDT analysis
}
