% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtEst.R
\name{sdtEst}
\alias{sdtEst}
\title{Calculate four SDT parameters at once: d', criterion, beta, and beta^2}
\usage{
sdtEst(h, fa)
}
\arguments{
\item{h}{The hit rate as a proportion.}

\item{fa}{The false alarm rate as a proportion.}
}
\value{
A dataframe with the hit rate, false alarm rate, and SDT values.
}
\description{
Calculate four SDT parameters at once: d', criterion, beta, and beta^2
}
\examples{
sdtEst(.72, .44)

}
